Fix 5 staging/intermediate errors from first prod run
- stg_lsq__lead_data: remove mx_sales_process (not in source view)
- stg_lsq__leads_test: null out product_type (not in source table)
- stg_superleap__opportunities: rename id -> opportunity_id
- stg_revenuecat__events: cast raw_payload::text for SUPER type compat
- int_marketing__payments_enriched: cast currency_conversion_rate::numeric
- dbt_project.yml: add Elementary flag for builtin materialization overrides

+++ models/intermediate/marketing/int_marketing__payments_enriched.sql
@@ -53,7 +53,7 @@ with_region as (
         end                                                                      as region,
 
         -- Amount in USD
-        round(amount * currency_conversion_rate, 2)                              as amount_usd,
+        round(amount::numeric * currency_conversion_rate::numeric, 2)            as amount_usd,
 
         -- Gross vs net booking
         case
